refactor: useEffect로 name input에 value 넣어주던거 onSuccess로 하도록 수정
